% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cutGroups}
\alias{cutGroups}
\title{split survival data into High and Low groups based on cutoff value and scores/expression values}
\usage{
cutGroups(os, out)
}
\arguments{
\item{os}{tibble with columns: samples, OS values, and Vital.Status}

\item{out}{output from mCut or mSet. It contains the cutoff threshold and either GSEA scores ("scores") or Gene Expression ("geneExp") values for each sample}
}
\value{
the os\link{"group"} column with samples split into High and Low groups
}
\description{
split survival data into High and Low groups based on cutoff value and scores/expression values
}
\examples{
os = dplyr::tibble(sample = letters[1:10], OS = 10:19, Vital.Status = c(rep("Alive",8),rep("Dead",2)))
out = list("cutoff" = 0.44, "scores" = rnorm(10))
os["group"] = cut(os,out)
os
}
