% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findENS}
\alias{findENS}
\title{convert gene symbols to ENSMBL ids}
\usage{
findENS(genes, ref)
}
\arguments{
\item{genes}{gene symbols}

\item{ref}{reference tibble with ensmble ids and gene symbols as columns}
}
\value{
character vector of ENSMBL ids for genes
}
\description{
convert gene symbols to ENSMBL ids
}
\examples{
genes = c("gene3","gene4","gene10")
ref = dplyr::tibble(ensg = paste0(rep("ENSG",10), 1:10), symbol = paste0(rep("gene",10), 10:1))
findENS(genes, ref)
}
