% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mCut}
\alias{mCut}
\title{provide cutoff threshold and gene expression values for specific gene}
\usage{
mCut(gene, mat, os, cut = "median")
}
\arguments{
\item{gene}{the gene we want to use to classify samples based on expression}

\item{mat}{tibble with samples and gene expression (FPKM)}

\item{os}{tibble with samples, overall survival (OS) values, and Vital.Status values}

\item{cut}{method to use to separate into high and low groups, can be either median (defualt) or cutP}
}
\value{
the cutoff threshold and expression values (log2(FPKM+1)) as a list
}
\description{
provide cutoff threshold and gene expression values for specific gene
}
\examples{
expr = dplyr::tibble(symbol = c("gene1","gene2","gene3"), A = c(0.063, 0.0213, 0.0059), B = c(0,0.1,0.5), C = c(0.03,0.05,0.22))
os = dplyr::tibble(sample = c("A","B","C"), OS = c(20.1,100.2,5.02), Vital.Status = c("Alive","Alive","Dead"))
mCut("gene1", mat = expr, os = os, cut = "cutP")
}
