% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mSet}
\alias{mSet}
\title{provide cutoff threshold and GSEA values for gene set of two or more genes}
\usage{
mSet(gs, ENS, os, ref, cut = "median")
}
\arguments{
\item{gs}{the genes we want to use to classify samples based on expression}

\item{ENS}{expression values (FPKM) tibble with samples as columns and ENSMBL ids as rows}

\item{os}{tibble with samples, overall survival (OS) values, and Vital.Status values}

\item{ref}{reference tibble with ensmble ids and gene symbols as columns, converts to ENS using this}

\item{cut}{method to use to separate into high and low groups, can be either median (defualt) or cutP}
}
\value{
the cutoff threshold and GSEA scores as a list
}
\description{
provide cutoff threshold and GSEA values for gene set of two or more genes
}
\examples{
expr = abs(matrix(rnorm(100),ncol=10))
rownames(expr) = paste0(rep("ENSG",10),1:10)
colnames(expr) = letters[1:10]
ref = dplyr::tibble(ensg = paste0(rep("ENSG",10), 1:10), symbol = paste0(rep("gene",10), 10:1))
os = dplyr::tibble(sample = letters[1:10], OS = 10:19, Vital.Status = c(rep("Alive",8),rep("Dead",2)))
mSet(c("gene1","gene4","gene5"), ENS = expr, os = os, ref = ref, cut = "cutP")
}
