% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotSurv}
\alias{plotSurv}
\title{plot the survival Kaplan-Meier curve}
\usage{
plotSurv(os, name = "gene")
}
\arguments{
\item{os}{overall survival tibble with High/Low groups formed}

\item{name}{name of the gene/geneSet for the title, default is "gene"}
}
\value{
plotting object
}
\description{
plot the survival Kaplan-Meier curve
}
\examples{
os = dplyr::tibble(sample = letters[1:10], OS = 10:19, Vital.Status = c(rep("Alive",8),rep("Dead",2)))
out = list("cutoff" = 0.44, "scores" = rnorm(10))
os["group"] = cut(os,out)
plotSurv(os, "gene1")
}
