% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{survStats}
\alias{survStats}
\title{calculate survival statistics for split survival data}
\usage{
survStats(os)
}
\arguments{
\item{os}{tibble with sample, OS, Vital.Status, and group (High/Low) columns}
}
\value{
stats tibble filled with test statistics for survival data: Comparison, survdiffP, coxHR, coxP
}
\description{
calculate survival statistics for split survival data
}
