% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mCut}
\alias{mCut}
\title{provide cutoff threshold and gene expression values for specific gene}
\usage{
mCut(gene, mat, os, cut = "median")
}
\arguments{
\item{gene}{the gene we want to use to classify samples based on expression}

\item{mat}{tibble with samples and gene expression (FPKM)}

\item{os}{tibble with samples, overall survival (OS) values, and Vital.Status values}

\item{cut}{method to use to separate into high and low groups, can be either median (defualt) or cutP}
}
\value{
the cutoff threshold and expression values (log2(FPKM+1)) as a list
}
\description{
provide cutoff threshold and gene expression values for specific gene
}
\examples{
data = data.frame(data = matrix(rnorm(100),ncol=10))
expr = dplyr::tibble(symbol = paste0(rep("gene",10),1:10), data)
colnames(expr) = c("symbol", letters[1:10])
os = dplyr::tibble(sample = letters[1:10], OS = 10:19, Vital.Status = c(rep("Alive",8),rep("Dead",2)))
mCut("gene1", mat = expr, os = os, cut = "cutP")
}
